refactor hooks and add new bgi section
