+++ static/templates/actors/npc-sheet.html
@@ -36,7 +36,7 @@
     <div class="npc-notes-container">
       <span class="item-title-npc">{{localize "TWODSIX.Actor.Notes.Notes"}}</span>
       {{#unless limited}}
-      <div contenteditable="true" data-edit="system.notes" style="height: 90%; border-radius: 0 0 2ch 2ch !important; overflow-y: auto">{{{system.notes}}}</div>
+      <div contenteditable="true" data-edit="system.notes" style="height: 90%; border-radius: 0 0 2ch 2ch !important; overflow-y: auto; padding-left: 2px;">{{{system.notes}}}</div>
      {{/unless}}
     </div>
     <div class="npc-skills">
